[test] Fix tests (#514)

+++ test/quantization/algorithm/test_gptq.py
@@ -147,7 +147,9 @@ class GPTQTest(unittest.TestCase):
         from transformers import AutoModelForCausalLM, AutoTokenizer
 
         tokenizer = AutoTokenizer.from_pretrained("Maykeye/TinyLLama-v0")
-        model = AutoModelForCausalLM.from_pretrained("Maykeye/TinyLLama-v0")
+        model = AutoModelForCausalLM.from_pretrained(
+            "Maykeye/TinyLLama-v0", dtype=torch.float32
+        )
 
         # Load data
         dataset = load_dataset("wikiText", "wikitext-2-raw-v1", split="train")
